requirements.txt
tensorflow==2.13.0
numpy==1.24.3
matplotlib==3.7.1
jupyter==1.0.0
Pillow>=10.3.0
streamlit>=1.37.0
protobuf==4.23.3
tensorflow-hub==0.14.0